requirements.txt
python-docx
pypiwin32==223